GripNet Integration
For integraing GripNet into embed module and example bug fix.
