+++ docs/source/kale.embed.rst
@@ -14,6 +14,27 @@ kale.embed.attention\_cnn module
    :undoc-members:
    :show-inheritance:
 
+kale.embed.gcn module
+----------------------------
+
+.. automodule:: kale.embed.gcn
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: message, update
+   :member-order: bysource
+
+kale.embed.gripnet module
+----------------------------
+
+.. automodule:: kale.embed.gripnet
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :exclude-members: message, update
+   :member-order: bysource
+
+
 kale.embed.image\_cnn module
 ----------------------------
