+++ examples/drug_gripnet/README.md
@@ -1,0 +1,8 @@
+# Link Classification: GripNet framework on PoSE sample dataset.
+
+## 1. Description
+This example is constructed by [GripNet Framework](https://github.com/NYXFLOWER/GripNet>) on Polypharmacy Side Effect (PoSE) prediction task. The PoSE sample dataset is from BioSNAP-Decagon dataset and the target is to predict relations between drug nodes on the whole gene-drug heterogeneous graph.
+
+
+## 2. Usage
+`python main.py --cfg configs/Drug_MINI-GripNet.yaml`
