+++ examples/drug_gripnet/configs/PoSE_MINI-GripNet.yaml
